// packages/api/src/generated/admin/models/adminRecruitmentPostSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */
import type { AdminRecruitmentPostSummaryResponsePositionsItem } from './adminRecruitmentPostSummaryResponsePositionsItem';
import type { AdminRecruitmentPostSummaryResponseProgressMethod } from './adminRecruitmentPostSummaryResponseProgressMethod';
import type { AdminRecruitmentPostSummaryResponseRecruitmentStatus } from './adminRecruitmentPostSummaryResponseRecruitmentStatus';
import type { AdminRecruitmentPostSummaryResponseRecruitmentType } from './adminRecruitmentPostSummaryResponseRecruitmentType';
import type { AdminRecruitmentPostSummaryResponseVisibility } from './adminRecruitmentPostSummaryResponseVisibility';

export interface AdminRecruitmentPostSummaryResponse {
  id: number;
  title: string;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `SIDE_PROJECT` | 1 | 사이드 프로젝트 |
     * | `STUDY` | 2 | 스터디 |
     */
  recruitmentType: AdminRecruitmentPostSummaryResponseRecruitmentType;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `ONLINE` | 1 | 온라인 |
     * | `OFFLINE` | 2 | 오프라인 |
     * | `HYBRID` | 3 | 온·오프라인 |
     */
  progressMethod: AdminRecruitmentPostSummaryResponseProgressMethod;
  capacity: number;
  /** 활동 기간(개월) */
  activityDurationMonths: number;
  positions: AdminRecruitmentPostSummaryResponsePositionsItem[];
  technologyStacks: string[];
  recruitmentStartDate: string;
  recruitmentEndDate: string;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `RECRUITING` | 1 | 모집 중 |
     * | `CLOSED` | 2 | 마감 |
     */
  recruitmentStatus: AdminRecruitmentPostSummaryResponseRecruitmentStatus;
  closedAt?: string;
  viewCount: number;
  bookmarkCount: number;
  commentCount: number;
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `VISIBLE` | 1 | 노출 |
     * | `HIDDEN` | 2 | 비노출 |
     */
  visibility: AdminRecruitmentPostSummaryResponseVisibility;
  authorUserId: number;
  /** 작성자 프로필이 없으면 null입니다. */
  authorNickname?: string;
  registeredAt: string;
}