// packages/api/src/generated/admin/models/changeAdminRecruitmentPostVisibilityRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */
import type { ChangeAdminRecruitmentPostVisibilityRequestVisibility } from './changeAdminRecruitmentPostVisibilityRequestVisibility';

export interface ChangeAdminRecruitmentPostVisibilityRequest {
  /**
     * @minItems 1
     * @maxItems 1000
     */
  ids: number[];
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `VISIBLE` | 1 | 노출 |
     * | `HIDDEN` | 2 | 비노출 |
     */
  visibility: ChangeAdminRecruitmentPostVisibilityRequestVisibility;
}