// packages/api/src/generated/admin/models/pageResponseAdminBootcampSummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */
import type { AdminBootcampSummaryResponse } from './adminBootcampSummaryResponse';
import type { PageInfo } from './pageInfo';

export interface PageResponseAdminBootcampSummaryResponse {
  items: AdminBootcampSummaryResponse[];
  pageInfo: PageInfo;
}